recent

+++ jc.h
@@ -4,4 +4,4 @@ int prt_str( char*);
 
 unsigned int bin_sum( unsigned int, unsigned int);
 
-float sumup( int);
+int sumup( int);
